% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_col.R
\name{define_col}
\alias{define_col}
\title{Define column types for data reading}
\usage{
define_col(
  ins = c("sf", "lf", "bsid"),
  adm = c("fixed", "adaptive", "tan", "pak", "ban", "bra", "cdi", "nld"),
  date_format = "",
  datetime_format = ""
)
}
\arguments{
\item{ins}{Character, the instrument}

\item{adm}{Character, fixed or adaptive}

\item{date_format}{Specification how to read the date field}

\item{datetime_format}{Specification how to read the datetime field}
}
\description{
Define column types for data reading
}
