% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_col_type.R
\name{define_col_type}
\alias{define_col_type}
\title{Define column types for data reading}
\usage{
define_col_type(
  ins = c("sf", "lf"),
  form = c("fixed", "adaptive"),
  date_format = "",
  datetime_format = ""
)
}
\arguments{
\item{ins}{Character, the instrument}

\item{form}{Character, fixed or adaptive}

\item{date_format}{Specification how to read the date field}

\item{datetime_format}{Specification how to read the datetime field}
}
\description{
Define column types for data reading
}
