% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bsid.R
\name{read_bsid}
\alias{read_bsid}
\title{Read data from Bayley Scales of Infant Development III}
\usage{
read_bsid(
  onedrive = Sys.getenv("ONEDRIVE_GSED"),
  path = "GSED Final Collated Phase 1 Data Files 18_05_22",
  verbose = FALSE,
  progress = FALSE,
  warnings = FALSE
)
}
\arguments{
\item{onedrive}{Character, the location of the local OneDrive sync}

\item{path}{Character, path name within the OneDrive}

\item{verbose}{Logical. Print diagnostic info? Defaults to `FALSE`.}

\item{progress}{Logical. Show progress in interactive session. Default to `FALSE`.}

\item{warnings}{Logical. Show warnings generated by `readr::read_csv()`.
Defaults to `FALSE`.}
}
\value{
A tibble with the original data and two column names: `file`
(containing the original file name) and `adm` (fixed or adaptive).
}
\description{
Read data from Bayley Scales of Infant Development III
}
