% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf.R
\name{read_sf}
\alias{read_sf}
\title{Read sf data}
\usage{
read_sf(
  adm = c("fixed", "adaptive"),
  onedrive = Sys.getenv("ONEDRIVE_GSED"),
  path = NULL,
  verbose = FALSE,
  progress = FALSE,
  warnings = FALSE
)
}
\arguments{
\item{adm}{Character, either "fixed" or "adaptive" administration type}

\item{onedrive}{Character, the location of the local OneDrive sync}

\item{path}{Character, path name within the OneDrive}

\item{verbose}{Logical. Print diagnostic info? Defaults to `FALSE`.}

\item{progress}{Logical. Show progress in interactive session. Default to `FALSE`.}

\item{warnings}{Logical. Show warnings generated by `readr::read_csv()`.
Defaults to `FALSE`.}
}
\value{
A tibble with the original data and two column names: `file`
(containing the original file name) and `adm` (fixed or adaptive).
If `adm == "fixed"` a tibble with 159
columns, with one test administration per row. If `adm == "adaptive"`, a tibble
with 14 columns, with one item administration per row.
}
\description{
Read sf data
}
