% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_variables.R
\name{rename_variables}
\alias{rename_variables}
\title{Rename variables}
\usage{
rename_variables(
  from,
  trans = c("original", "sequence", "connect", "extend"),
  lowercase = TRUE,
  underscore = TRUE,
  trim = TRUE
)
}
\arguments{
\item{from}{A character vector with column names}

\item{trans}{One of "original", "sequence", "connect" or "extend". Desired
transformation of `from`. "original" and "sequence" preserve the original item
names or sequence. "connect" and "extend" translate items names to the 9-position
GSED lexicon. "connect" uses existing item names (useful for joint analysis with
existing data), whereas "extend" defines new 9-position names for the item names
in `from`.}

\item{lowercase}{Sets all variables in lower case.}

\item{underscore}{Replaces space (" ") and dash ("-") by underscore ("_")}

\item{trim}{Removes the "Ma_" string}
}
\value{
A character vector of the same length as `from` with processed
names.
}
\description{
Rename variables
}
\examples{
data <- read_sf()
from <- colnames(data)
idx <- c(1:3, 19, 21:23)
from[idx]
to <- rename_variables(from, "sequence", lowercase = FALSE)
to[idx]
to <- rename_variables(from, "extend")
to[idx]
}
