% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_visits.R
\name{repair_visits}
\alias{repair_visits}
\title{Repair and standardize visit data}
\usage{
repair_visits(visits, quiet = FALSE)
}
\arguments{
\item{visits}{A data frame containing visit data with required columns.}

\item{quiet}{Logical, if \code{TRUE}, suppresses output messages.}
}
\value{
A data frame with repaired and standardized visit data.
}
\description{
This function repairs and standardizes \code{visit} data after being read by
\code{read_gsed_fixed()} to ensure that the \code{visit} data is in a consistent
format.
}
\details{
The function performs several operations:
\itemize{
\item Ensuring required columns are present
\item Repairing \code{subjid} according to ISO 3166-1 numeric code
\item Defining GSED cohort and cohort number
\item Removing duplicate visits based on unique identifiers
\item Sorting the data by \code{cohort}, \code{subjid}, \code{agedays}, \code{vist_type}, and \code{ins}
}

The function is to be run after \code{read_gsed_fixed()}

A unique visit identifier consists of: \code{subjid}, \code{agedays}, \code{vist_type}, \code{ins}
}
