% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bsid.R
\name{read_bsid}
\alias{read_bsid}
\title{Read data from Bayley Scales of Infant Development III}
\usage{
read_bsid(
  onedrive = Sys.getenv("ONEDRIVE_GSED"),
  path = NULL,
  verbose = FALSE,
  progress = FALSE,
  warnings = FALSE
)
}
\arguments{
\item{onedrive}{Character, the location of the local OneDrive sync}

\item{path}{Character, path name within the OneDrive}

\item{verbose}{Logical. Print diagnostic info? Defaults to \code{FALSE}.}

\item{progress}{Logical. Show progress in interactive session. Default to \code{FALSE}.}

\item{warnings}{Logical. Show warnings generated by \code{readr::read_csv()}.
Defaults to \code{FALSE}.}
}
\value{
A tibble with the original data and two column names: \code{file}
(containing the original file name) and \code{adm} (fixed or adaptive).
}
\description{
Read data from Bayley Scales of Infant Development III
}
\examples{
onedrive <- Sys.getenv("ONEDRIVE_GSED")

}
