% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf.R
\name{read_sf}
\alias{read_sf}
\title{Read sf data}
\usage{
read_sf(
  type = c("fixed", "adaptive"),
  onedrive = Sys.getenv("ONEDRIVE_GSED"),
  path = "GSED Final Collated Phase 1 Data Files 18_05_22"
)
}
\arguments{
\item{type}{Character, either "fixed" or "adaptive"}

\item{onedrive}{Character, the location of the local OneDrive sync}

\item{path}{Character, path name within the OneDrive}
}
\value{
A tibble with the original data and one new column named `file`
containing the original file name. If `type == "fixed"` a tibble with 159
columns, with one test administration per row. If `type == "adaptive"`, a tibble
with 14 columns, with one item administration per row.
}
\description{
Read sf data
}
